Add CUDA toolkit support with multiple CUDA versions
Adds CUDA 12.2.2, 12.4.1, 12.6.3, 12.9.1 and 13.0.0 downloads,
nvvm/bin to PATH, PRESERVE_PKG flag to keep bin/ directory,
and fixes ffnvcodec version selection.

+++ util/run_stage.sh
@@ -32,7 +32,8 @@ if [[ "$SELF" == */??-*/??-*.sh && -d "$FFBUILD_DESTDIR" ]]; then
     cp -al "$FFBUILD_DESTDIR"/. /
 fi
 
-rm -rf "$FFBUILD_DESTPREFIX"/bin
+# Only clean bin/ if not explicitly needed
+[[ "$PRESERVE_PKG" != "1" ]] && rm -rf "$FFBUILD_DESTPREFIX"/bin
 
 if [[ -n "$STAGENAME" ]]; then
     rm -rf "/$STAGENAME"
